FinDecide/requirements.txt
transformers
pandas
torch
streamlit
python-dotenv
PyPDF2
langchain-community
langchain
openai
faiss-cpu
altair
tiktoken
huggingface_hub
pyth
langchain_openai